Test #2 - Controller Test TDD 방식을 통하여 나머지 컨트롤러에 대해 테스트 -> 구현의 순서로 테스트주도 개발 연습

+++ src/main/java/com/lee/placein/controller/EventController.java
@@ -14,7 +14,7 @@ public class EventController {
     }
 
     @GetMapping("/{eventId}")
-    public String eventDetail(@PathVariable Integer eventId){
+    public String eventDetail(@PathVariable Long eventId){
         return "event/detail";
     }
 }
